Rename api-client.
